Added support for inner/base http client in Firestore constructors. (#5)
Added support for inner/base http client in Firestore constructors.

You can now pass `Firestore` an `package:http` `Client` which can be used to add logging or request modification.  

For example:   
```dart
final firestore = Firestore.newInstance(
  credentials: myCredentials,
  innerClient: LoggingHttpClient(),
);

class LoggingHttpClient extends http.BaseClient {
  final http.Client _delegate = http.Client();

  @override
  Future<http.StreamedResponse> send(http.BaseRequest request) {
    print('Sending request: $request');
    return _delegate.send(request);
  }
}
```

Additionally I added a `run-remote-tests` workflow that runs all tests that call the remote Firestore server.  
Another workflow running the other tests using the emulator needs to be added as well in the future.

This is also a prerequisite to be able to make a regression test for #4.